+++ Constraid/FlushWithEdges.swift
@@ -6,7 +6,7 @@
 
 extension ConstraidView {
     @discardableResult
-    public func flush(withLeadingEdgeOf item: Any?, constant: CGFloat = 0.0,
+    public func flush(withLeadingEdgeOf item: Any?, insetBy inset: CGFloat = 0.0,
                     multiplier: CGFloat = 1.0,
                     priority: ConstraidLayoutPriority = ConstraidLayoutPriorityRequired) -> ConstraidConstraintCollection {
 
@@ -14,14 +14,14 @@ extension ConstraidView {
         let collection = ConstraidConstraintCollection([
             NSLayoutConstraint(item: self, attribute: .leading, relatedBy: .equal,
                                toItem: item, attribute: .leading, multiplier: multiplier,
-                               constant: constant, priority: priority)
+                               constant: inset, priority: priority)
             ])
         collection.activate()
         return collection
     }
 
     @discardableResult
-    public func flush(withTrailingEdgeOf item: Any?, constant: CGFloat = 0.0,
+    public func flush(withTrailingEdgeOf item: Any?, insetBy inset: CGFloat = 0.0,
                     multiplier: CGFloat = 1.0,
                     priority: ConstraidLayoutPriority = ConstraidLayoutPriorityRequired) -> ConstraidConstraintCollection {
 
@@ -29,14 +29,14 @@ extension ConstraidView {
         let collection = ConstraidConstraintCollection([
             NSLayoutConstraint(item: self, attribute: .trailing, relatedBy: .equal,
                                toItem: item, attribute: .trailing, multiplier: multiplier,
-                               constant: constant, priority: priority)
+                               constant: inset, priority: priority)
             ])
         collection.activate()
         return collection
     }
 
     @discardableResult
-    public func flush(withTopEdgeOf item: Any?, constant: CGFloat = 0.0,
+    public func flush(withTopEdgeOf item: Any?, insetBy inset: CGFloat = 0.0,
                     multiplier: CGFloat = 1.0,
                     priority: ConstraidLayoutPriority = ConstraidLayoutPriorityRequired) -> ConstraidConstraintCollection {
 
@@ -44,7 +44,7 @@ extension ConstraidView {
         let constraints = ConstraidConstraintCollection([
             NSLayoutConstraint(item: self, attribute: .top, relatedBy: .equal,
                                toItem: item, attribute: .top, multiplier: multiplier,
-                               constant: constant, priority: priority)
+                               constant: inset, priority: priority)
             ])
         constraints.activate()
         return constraints
@@ -52,7 +52,7 @@ extension ConstraidView {
 
 
     @discardableResult
-    public func flush(withBottomEdgeOf item: Any?, constant: CGFloat = 0.0,
+    public func flush(withBottomEdgeOf item: Any?, insetBy inset: CGFloat = 0.0,
                     multiplier: CGFloat = 1.0,
                     priority: ConstraidLayoutPriority = ConstraidLayoutPriorityRequired) -> ConstraidConstraintCollection {
 
@@ -60,45 +60,45 @@ extension ConstraidView {
         let constraints = ConstraidConstraintCollection([
             NSLayoutConstraint(item: self, attribute: .bottom, relatedBy: .equal,
                                toItem: item, attribute: .bottom, multiplier: multiplier,
-                               constant: constant, priority: priority)
+                               constant: inset, priority: priority)
             ])
         constraints.activate()
         return constraints
     }
 
     @discardableResult
-    public func flush(withVerticalEdgesOf item: Any?, constant: CGFloat = 0.0,
+    public func flush(withVerticalEdgesOf item: Any?, insetBy inset: CGFloat = 0.0,
                     multiplier: CGFloat = 1.0,
                     priority: ConstraidLayoutPriority = ConstraidLayoutPriorityRequired) -> ConstraidConstraintCollection {
 
-        let constraints = flush(withLeadingEdgeOf: item, constant: constant, multiplier: multiplier,
+        let constraints = flush(withLeadingEdgeOf: item, insetBy: inset, multiplier: multiplier,
                                 priority: priority) +
-                          flush(withTrailingEdgeOf: item, constant: (-1.0 * constant),
+                          flush(withTrailingEdgeOf: item, insetBy: (-1.0 * inset),
                                 multiplier: multiplier, priority: priority)
         constraints.activate()
         return constraints
     }
 
     @discardableResult
-    public func flush(withHorizontalEdgesOf item: Any?, constant: CGFloat = 0.0,
+    public func flush(withHorizontalEdgesOf item: Any?, insetBy inset: CGFloat = 0.0,
                     multiplier: CGFloat = 1.0,
                     priority: ConstraidLayoutPriority = ConstraidLayoutPriorityRequired) -> ConstraidConstraintCollection {
 
-        let constraints = flush(withTopEdgeOf: item, constant: constant, multiplier: multiplier,
+        let constraints = flush(withTopEdgeOf: item, insetBy: inset, multiplier: multiplier,
                                 priority: priority) +
-                          flush(withBottomEdgeOf: item, constant: (-1.0 * constant),
+                          flush(withBottomEdgeOf: item, insetBy: (-1.0 * inset),
                                 multiplier: multiplier, priority: priority)
         constraints.activate()
         return constraints
     }
 
     @discardableResult
-    public func flush(withEdgesOf item: Any?, constant: CGFloat = 0.0, multiplier: CGFloat = 1.0,
+    public func flush(withEdgesOf item: Any?, insetBy inset: CGFloat = 0.0, multiplier: CGFloat = 1.0,
                     priority: ConstraidLayoutPriority = ConstraidLayoutPriorityRequired) -> ConstraidConstraintCollection {
         
-        let constraints = flush(withHorizontalEdgesOf: item, constant: constant,
+        let constraints = flush(withHorizontalEdgesOf: item, insetBy: inset,
                                 multiplier: multiplier, priority: priority) +
-                          flush(withVerticalEdgesOf: item, constant: constant,
+                          flush(withVerticalEdgesOf: item, insetBy: inset,
                                 multiplier: multiplier, priority: priority)
         constraints.activate()
         return constraints